% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_name}
\alias{clean_name}
\title{Clean and standardize LEA/school names}
\usage{
clean_name(x)
}
\arguments{
\item{x}{Character vector of names}
}
\value{
Cleaned character vector
}
\description{
Clean and standardize LEA/school names
}
\keyword{internal}
