% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{create_empty_assessment_processed}
\alias{create_empty_assessment_processed}
\title{Create empty processed assessment data frame}
\usage{
create_empty_assessment_processed()
}
\value{
Empty data frame with processed assessment columns
}
\description{
Create empty processed assessment data frame
}
\keyword{internal}
