% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{download_assessment_data}
\alias{download_assessment_data}
\title{Download assessment data from NC DPI}
\usage{
download_assessment_data()
}
\value{
Data frame with all years of assessment data
}
\description{
Downloads the School Report Cards data set which contains assessment
performance counts. The file is a large text file (~900MB uncompressed)
inside a zip archive.
}
\keyword{internal}
