% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_ccd_file}
\alias{download_ccd_file}
\title{Download a CCD data file}
\usage{
download_ccd_file(end_year, level)
}
\arguments{
\item{end_year}{School year end}

\item{level}{"lea" or "school"}
}
\value{
Data frame filtered to North Carolina
}
\description{
Download a CCD data file
}
\keyword{internal}
