% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_nc_stat_profile}
\alias{download_nc_stat_profile}
\title{Download from NC DPI Statistical Profile}
\usage{
download_nc_stat_profile(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
List with lea and school data frames
}
\description{
Downloads enrollment data from NC DPI's Oracle APEX REST API.
Uses the Statistical Profile interactive reports which support CSV export.
}
\keyword{internal}
