% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_nces_ccd}
\alias{download_nces_ccd}
\title{Download from NCES CCD}
\usage{
download_nces_ccd(end_year)
}
\arguments{
\item{end_year}{School year end}
}
\value{
List with lea and school data frames
}
\description{
Downloads enrollment data from NCES Common Core of Data.
CCD provides standardized school data for all 50 states.
}
\keyword{internal}
