% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{download_stat_profile_table}
\alias{download_stat_profile_table}
\title{Download a specific table from NC DPI Statistical Profile}
\usage{
download_stat_profile_table(end_year, table_type)
}
\arguments{
\item{end_year}{School year end}

\item{table_type}{Type of table: "lea_enrollment", "lea_race", "school_enrollment"}
}
\value{
Data frame
}
\description{
Download a specific table from NC DPI Statistical Profile
}
\keyword{internal}
