% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_assessment_data_url}
\alias{get_assessment_data_url}
\title{Get assessment data URL}
\usage{
get_assessment_data_url()
}
\value{
URL string
}
\description{
Returns the URL for downloading assessment data from NC DPI.
}
\keyword{internal}
