% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_assessment.R
\name{get_available_assessment_years}
\alias{get_available_assessment_years}
\title{Get available assessment years}
\usage{
get_available_assessment_years()
}
\value{
A list with:
\item{years}{Vector of available years}
\item{note}{Description of data availability}
}
\description{
Returns the range of years for which assessment data is available
from the NC DPI School Report Cards data.
}
\examples{
years <- get_available_assessment_years()
print(years$years)
}
