% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_directory.R
\name{get_raw_directory}
\alias{get_raw_directory}
\title{Download raw directory data from NCDPI}
\usage{
get_raw_directory(directory_type = "all")
}
\arguments{
\item{directory_type}{Type of directory ("private_schools", "all")}
}
\value{
List with raw data frames for each directory type requested
}
\description{
Downloads the North Carolina school directory Excel file(s).
}
\keyword{internal}
