% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from NC DPI}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2023-24 = 2024)}
}
\value{
List with lea and school data frames
}
\description{
Downloads LEA and school enrollment data from NC DPI's Statistical Profile
system.
}
\keyword{internal}
