% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_assessment.R
\name{process_assessment}
\alias{process_assessment}
\title{Process raw NC assessment data}
\usage{
process_assessment(raw_data, end_year)
}
\arguments{
\item{raw_data}{Data frame from get_raw_assessment}

\item{end_year}{School year end}
}
\value{
Processed data frame with standardized columns
}
\description{
Transforms raw data from NC DPI into a standardized schema with
clear column names and identifiers.
}
\keyword{internal}
