% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_enr}
\alias{process_enr}
\title{Process raw NC enrollment data}
\usage{
process_enr(raw_data, end_year)
}
\arguments{
\item{raw_data}{List containing lea and school data frames from get_raw_enr}

\item{end_year}{School year end}
}
\value{
Processed data frame with standardized columns
}
\description{
Transforms raw data from NC DPI into a standardized schema
combining LEA and school data.
}
\keyword{internal}
