% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_enrollment.R
\name{process_lea_enr}
\alias{process_lea_enr}
\title{Process LEA-level (district) enrollment data}
\usage{
process_lea_enr(df, end_year)
}
\arguments{
\item{df}{Raw LEA data frame}

\item{end_year}{School year end}
}
\value{
Processed LEA data frame
}
\description{
Process LEA-level (district) enrollment data
}
\keyword{internal}
