% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_year}
\alias{validate_year}
\title{Validate school year}
\usage{
validate_year(end_year, min_year = 2006, max_year = 2025)
}
\arguments{
\item{end_year}{School year end}

\item{min_year}{Minimum valid year}

\item{max_year}{Maximum valid year (currently 2025)}
}
\value{
TRUE if valid, otherwise throws error
}
\description{
Validate school year
}
\keyword{internal}
