% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_available_years}
\alias{get_available_years}
\title{Get available years}
\usage{
get_available_years()
}
\value{
Character vector of available years
}
\description{
Returns the range of years for which enrollment data is available.
}
\examples{
get_available_years()
}
